server/requirements.txt
Django==2.1
djangorestframework==3.9.1
psycopg2-binary
django-import-export==1.2.0
django-rest-swagger==2.2.0
Pillow==6.2.0
PyJWT==1.7.1
django-cors-headers==2.2.0
django_extensions
python-crontab
django-redis
redis
pyvmomi==6.7.3
suds-jurko
vcrpy>=1.1.1
# test and cov
coverage==4.5.2
pytest
pytest-django==3.7.0
pytest-cov
coverage-badge
# lint
pylint==2.3.0
pylint-django==2.0.6
pylint-plugin-utils==0.5 
